refactor: 일급 컬렉션 Cars 추가
