+++ src/main/java/racing/Cars.java
@@ -1,0 +1,21 @@
+package racing;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class Cars {
+	private List<Car> cars;
+
+	Cars(String namesWithDelimiter) {
+		this.cars = new ArrayList<>();
+
+		CarNames carNames = new CarNames(namesWithDelimiter);
+		for (CarName carName : carNames.getNames()) {
+			this.cars.add(new Car(carName));
+		}
+	}
+
+	public List<Car> getCars() {
+		return this.cars;
+	}
+}
